# Make vm_call_cfunc_with_frame a fastpath