--- a/benchmark/vm_send_cfunc.yml
+++ b/benchmark/vm_send_cfunc.yml
@@ -1,0 +1,3 @@
+benchmark:
+  vm_send_cfunc: self.class
+loop_count: 100000000